Revert 'About' page text
I was originally given a document for the 'About' page text, which I
tried to modify for improved clarity and to represent the actual
organization of some pages. But I appear to have changed it too much.
Revert the text back to pretty much a literal copy-paste of the
original text, as requested.