Dynamically set the "site source" footer link
Currently, the "site source" link in the footer is always set to the
same thing. But if we're serving a "scratch"/staging version of the
site, the source may be different. Use the site.github.repository_url
var to link to the actual repository being used. This can be a helpful
reminder of which site we're actually looking at.