Replace site.github.url with site.url
In recent versions of github-pages, you can use site.url instead of
site.github.url. When previewing locally, site.url will automatically be
set to localhost and in production it will be set to your appropriate
Github URL.

Update the templates so a locally generated site can be previewed before
pushing to github.